xphoto: move inpainting FSR algorithm into a separate file
- get rid of cv::
